requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.1
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
zope.interface==5.4.0
djangorestframework
hazm
